verified for 2.2, make it backwards compat
