+++ tests/sentry/new_migrations/monkey/test_monkey.py
@@ -1,9 +1,9 @@
 from django.db.migrations import executor, writer
 from django.db.models import Field
 
+from sentry.new_migrations.monkey import SENTRY_MIGRATION_TEMPLATE
 from sentry.new_migrations.monkey.executor import SentryMigrationExecutor
 from sentry.new_migrations.monkey.fields import deconstruct, original_deconstruct
-from sentry.new_migrations.monkey.writer import SENTRY_MIGRATION_TEMPLATE
 from sentry.testutils import TestCase
